+++ src/main/java/com/sam/profilecreation_service/service/CoachService.java
@@ -8,7 +8,10 @@ public interface CoachService {
 
     CoachEntity createCoach(CoachEntity CoachEntity);
 
-    CoachEntity findCoachById(Integer id);
+    //CoachEntity findCoachById(Integer id);
+
+    CoachEntity findCoachById(Long id);
+
     List<CoachEntity> findAllCoaches();
     CoachEntity updateCoach(CoachEntity CoachEntity, Long id);
     void deleteCoach(Long id);
